# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Penjualan",
        "description": "API untuk manajemen data penjualan",
        "contact": {
            "name": "Muhammad Gibran",
            "email": "gibran.devv@email.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/penjualan/{id}": {
            "get": {
                "tags": [
                    "Penjualan"
                ],
                "summary": "Mendapatkan detail penjualan berdasarkan ID",
                "operationId": "aecc3b47fcaf3aee897807f0c23194bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID penjualan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan detail penjualan"
                    },
                    "404": {
                        "description": "Penjualan tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Penjualan"
                ],
                "summary": "Memperbarui data penjualan",
                "operationId": "2fa08c86459b832668bb61614f077882",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID penjualan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data penjualan yang diperbarui",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-05"
                                    },
                                    "id_pelanggan": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil memperbarui data penjualan"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Penjualan"
                ],
                "summary": "Menghapus penjualan",
                "operationId": "948a0e1923f1606eee3e3285ed288d78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID penjualan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menghapus penjualan"
                    },
                    "404": {
                        "description": "Penjualan tidak ditemukan"
                    }
                }
            }
        },
        "/api/penjualan/create": {
            "post": {
                "tags": [
                    "Penjualan"
                ],
                "summary": "Membuat penjualan baru",
                "operationId": "700252c0366540d84c24c86b8c729fdb",
                "requestBody": {
                    "description": "Data penjualan baru",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-04"
                                    },
                                    "id_pelanggan": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil membuat penjualan baru"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
...
    },
...
}